newton/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==36.0.1
cytoolz==0.11.2
Deprecated==1.2.13
Django==4.0.3
django-email-log==1.2.0
django-extensions==3.1.5
django-redis==5.2.0
django-storages==1.12.3
django-werkzeug-debugger-runserver==0.3.1
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
fastecdsa==2.2.3
frozenlist==1.3.0
hexbytes==0.2.2
idna==3.3
jmespath==0.10.0
jsonschema==3.2.0
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
packaging==21.3
parsimonious==0.8.1
phonenumbers==8.12.44
Pillow==9.0.1
protobuf==3.19.4
pycparser==2.21
pycryptodome==3.14.1
PyJWT==2.3.0
pyOpenSSL==22.0.0
pysha3==1.0.2
python-dateutil==2.8.2
redis==4.1.4
requests==2.27.1
rlp==2.0.1
sentry-sdk==1.5.6
six==1.16.0
sqlparse==0.4.2
toolz==0.11.2
urllib3==1.26.8
uWSGI==2.0.20
varint==1.0.2
web3==5.28.0
websockets==9.1
Werkzeug==2.0.3
wrapt==1.13.3
yarl==1.7.2
filetype==1.0.10
python-openid2